feat(rooms): move files to rooms folder
