+++ src/rooms/rooms.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { RoomsController } from './controllers/rooms.controller';
-import { RoomsService } from './services/rooms.service';
+import { RoomsController } from './controllers/rooms/rooms.controller';
+import { RoomsService } from './services/rooms/rooms.service';
 import { FloorsService } from './services/floors/floors.service';
 import { Floor } from './entities/floor.entity';
 import { FloorsController } from './controllers/floors/floors.controller';
